requirements.txt
python-dotenv==1.1.0
google-genai==0.8.0
